SDK-1975: Fixed modify time set

+++ src/Upgrade/Infrastructure/PackageManager/Reader/ComposerLockReader.php
@@ -69,7 +69,7 @@ class ComposerLockReader implements ComposerLockReaderInterface
         if (!file_exists($path)) {
             throw new FileNotFoundException('File is not exist: ' . $path);
         }
-        $fileTime = filemtime($path);
+        $fileTime = (int)filemtime($path);
         if (!$this->composerLockData || $fileTime > $this->modifyTime) {
             $this->modifyTime = $fileTime;
             $this->composerLockData = json_decode((string)file_get_contents($path), true);
